requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# HTTP Requests
requests==2.31.0

# AI Models
google-generativeai==0.8.3
openai==1.6.0

# PDF/Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1

# Environment
python-dotenv==1.0.0